@@ -1,5 +1,5 @@
 {
-  "name": "job-portal",
+  "name": "pac",
   "version": "0.0.0",
   "license": "MIT",
   "angular-cli": {},
@@ -15,36 +15,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^2.3.1",
-    "@angular/compiler": "^2.3.1",
-    "@angular/core": "^2.3.1",
-    "@angular/forms": "^2.3.1",
-    "@angular/http": "^2.3.1",
-    "@angular/platform-browser": "^2.3.1",
-    "@angular/platform-browser-dynamic": "^2.3.1",
-    "@angular/router": "^3.3.1",
+    "@angular/common": "^5.2.4",
+    "@angular/compiler": "^5.2.4",
+    "@angular/core": "^5.2.4",
+    "@angular/forms": "^5.2.4",
+    "@angular/http": "^5.2.4",
+    "@angular/platform-browser": "^5.2.4",
+    "@angular/platform-browser-dynamic": "^5.2.4",
+    "@angular/platform-server": "^5.2.4",
+    "@angular/router": "^5.2.4",
+    "@ngrx/store": "^5.0.0",
+    "body-parser": "^1.18.2",
     "bootstrap": "^3.3.7",
     "cookie-parser": "^1.4.3",
     "core-js": "^2.4.1",
     "dateformat": "^3.0.3",
     "express-session": "^1.15.6",
-    "jsonwebtoken": "^8.2.1",
-    "lodash": "^4.17.10",
+    "http": "0.0.0",
+    "jsonwebtoken": "^8.1.1",
+    "jspdf": "^1.3.5",
+    "lodash": "^4.17.5",
     "md5": "^2.2.1",
     "moment": "^2.22.1",
     "mongodb": "^3.1.0-beta4",
-    "mysql": "^2.15.0",
     "nexmo": "^2.2.1",
-    "ngx-dropdown": "0.0.22",
+    "ngx-filter-pipe": "^2.1.0",
     "otp-generator": "^1.1.0",
     "passport": "^0.4.0",
+    "passport-jwt": "^3.0.1",
     "passport-local": "^1.0.0",
+    "promise-polyfill": "7.1.0",
     "rxjs": "^5.0.1",
     "ts-helpers": "^1.1.1",
     "zone.js": "^0.7.2"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^2.3.1",
+    "@angular/compiler-cli": "^5.2.4",
     "@types/jasmine": "2.5.38",
     "@types/node": "^6.0.42",
     "angular-cli": "1.0.0-beta.28.3",
@@ -59,6 +65,6 @@
     "protractor": "~4.0.13",
     "ts-node": "1.2.1",
     "tslint": "^4.3.0",
-    "typescript": "~2.0.3"
+    "typescript": "^2.7.1"
   }
 }
