@@ -1,6 +1,6 @@
 import { CanActivate, CanLoad, ActivatedRouteSnapshot, RouterStateSnapshot, Router, Route, ActivatedRoute } from '@angular/router';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs/observable';
 import { AuthService } from './auth.service';
 
 @Injectable()
@@ -13,7 +13,7 @@ export class AuthGuard implements CanActivate, CanLoad {
     //    2.) If token is not set, which means a new tab is opened, then check the node server if session is still open. if open then take action similar to 1, ese redirect to sigin page
     //     console.log("Check for session");
         return this.authService.isSessionOpen().take(1).map(obj => {
-            console.log(obj);
+            //console.log(obj);
             if (obj == true) {
                 return obj;
             }
